// src/app/pages/home/pages/fornecedor/models/fornecedor.ts
export interface IFornecedor {
  id?: number;
  fantasia: string;
  razaoSocial: string;
  status: string;
  cpfCnpj: string;
  inscEstadual: string;
  cep: string;
  logradouro: string;
  numero: number;
  complemento: string
  bairro: string
  cidade: string
  uf: string
  email: string
  celular: string
  telefone: string
  fax: string
  pessoaContato: string
  observacoes: string
}